Buttons working (2)
The side buttons now fully working as do the menu buttons (except
settings since there are no settings). I adjusted the amount it slides
and added the little arrow to show that there was a side menu.  2 of the
side buttons close the menu while the other two do not. This was to show
the difference so we can decide which we like better.