2013 - 2014 update
Updated with information for 2013-2014. I also updated the string for
the version alert and added a please rate box text. In the MainActivity
file I added a simple way to update the attendance string file, changed
what classifies it as a special (a * in front of the time as well as add
times), added a please rate box after 5 launches, and added notes.